working on those last few test failures

+++ vivisect/analysis/arm/emulation.py
@@ -159,13 +159,13 @@ def buildFunctionApi(vw, fva, emu, emumon):
 
     if callconv == 'armcall':
         if emumon.stackmax > 0:
-            targc = (emumon.stackmax / 8) + 6
+            targc = int(emumon.stackmax / 8) + 6
             if targc > 40:
                 emumon.logAnomaly(emu, fva, 'Crazy Stack Offset Touched: 0x%.8x' % emumon.stackmax)
             else:
                 argc = targc
 
-        funcargs = [('int',archargname(i)) for i in range(argc)]
+        funcargs = [('int', archargname(i)) for i in range(argc)]
 
     api = ('int', None, callconv, None, funcargs)
     vw.setFunctionApi(fva, api)
